// src/models/regions/hubs/traits/types.ts
export type HubTrait =
  | 'ancient infrastructure'
  | 'black market'
  | 'buried ruins'
  | 'chaotic layout'
  | 'corrupt laws'
  | "coup d'etat"
  | 'cultural center'
  | 'defiled land'
  | 'despotic lord'
  | 'divine intervention'
  | 'draconian laws'
  | 'dueling warlords'
  | 'dying forest'
  | 'ethnic tensions'
  | 'faded remnant'
  | 'foreign enclave'
  | 'fortress monastery'
  | 'fungal garden'
  | 'fungal infestation'
  | 'gang warfare'
  | 'grand library'
  | 'great famine'
  | 'guild oligarchy'
  | 'harsh occupation'
  | 'heavy fortification'
  | 'heretical beliefs'
  | 'hidden ruler'
  | 'hostile terrain'
  | 'incompetent leaders'
  | 'invading army'
  | 'knightly order'
  | 'lawless chaos'
  | 'magical academy'
  | 'merchant council'
  | 'monstrous tribute'
  | 'mercenary companies'
  | 'natural disaster'
  | 'neglectful ruler'
  | 'new industry'
  | 'nomadic traders'
  | 'opulent court'
  | 'oppressed minority'
  | 'old wounds'
  | 'pilgrimage site'
  | 'pirate haven'
  | 'plague outbreak'
  | 'poacher problems'
  | 'population boom'
  | 'prison colony'
  | 'prosperous land'
  | 'punishment post'
  | 'raider scourge'
  | 'rebel stronghold'
  | 'recent brutality'
  | 'reclusive hermit'
  | 'reformist struggle'
  | 'regime change'
  | 'remnant faith'
  | 'resource depletion'
  | 'revanchist exile'
  | 'ritual combat'
  | 'ruined castle'
  | 'sacred grove'
  | 'secret police'
  | 'sectarian violence'
  | 'sinking city'
  | 'slave markets'
  | 'slave uprising'
  | 'struggling colony'
  | 'social stratification'
  | 'sorcerous cabal'
  | 'theocratic authority'
  | 'toxic economy'
  | 'syncretic faith'
  | 'unique product'
  | 'upcoming festival'
  | 'war conscripts'
  | 'witch hunts'
  | 'xenophobic locals'

export type HubTraitDetails = {
  text: string
  coastal?: boolean
  urban?: boolean
  warfare?: boolean
  conflicts?: HubTrait[]
}